+++ app/decorators/watercraft_storage_pass_decorator.rb
@@ -1,4 +1,4 @@
-class WatercraftStoragePassDecorator < Draper::Decorator
+class WatercraftStoragePassDecorator < AmenityPassDecorator
   delegate_all
 
   # Define presentation-specific methods here. Helpers are accessed through
@@ -10,7 +10,7 @@ class WatercraftStoragePassDecorator < Draper::Decorator
   #     end
   #   end
 
-  def property_summary(type: :icons)
-    h.render 'residencies/property_icon_list', residencies: object.resident.residencies.decorate
+  def self.icon_name
+    helpers.icon_for_scope('watercraft')
   end
 end